debug de l'impossibilite de parier : j'avai modifie le return de remaining_credits pour afficher les float avec deux chiffres apres la virgule et il kiffe pas le php

+++ db_install.php
@@ -27,7 +27,7 @@ $qTbMatch = "CREATE TABLE IF NOT EXISTS `match` (
   `date` date NOT NULL,
   `odds1` float(11) NOT NULL,  
   `odds2` float(11) NOT NULL,  
-  `draw` float NOT NULL, 
+  `draw` float(11) NOT NULL, 
   `res_type` int(11) NOT NULL,
   PRIMARY KEY (`ref`)
 ) ENGINE=InnoDB;";
